disclosureAnalysis/requirements.txt
beautifulsoup4>=4.12.2
lxml>=5.1.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0

# OpenTelemetry for logs/metrics export to central collector
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp>=1.28.0
opentelemetry-instrumentation-logging>=0.52b0
